# HW10 Program Optimization