--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -2,10 +2,14 @@ module github.com/AlexandrNikitin3776/OTUS2021Golang_Developer/hw10_program_opti
 
 go 1.19
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/mailru/easyjson v0.7.7
+	github.com/stretchr/testify v1.7.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/josharian/intern v1.0.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )