--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -3,7 +3,6 @@ package hw10programoptimization
 import (
 	"bufio"
 	"encoding/json"
-	"fmt"
 	"io"
 	"strings"
 )
@@ -19,46 +18,60 @@ type User struct {
 }
 
 type DomainStat map[string]int
-type users []User
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
+	errChannel := make(chan error)
+	defer close(errChannel)
+
+	usersChannel := getUsers(r, errChannel)
+	return countDomains(usersChannel, errChannel, domain)
 }
 
-func getUsers(r io.Reader) (users, error) {
-	bufferedReader := bufio.NewReader(r)
-	result := make(users, 0)
+func getUsers(r io.Reader, errChannel chan error) <-chan User {
+	usersChannel := make(chan User)
 
-	var user User
-	for {
-		line, err := bufferedReader.ReadBytes('\n')
-		if len(line) == 0 && err != nil {
-			if err == io.EOF {
-				return result, nil
+	go func() {
+		defer close(usersChannel)
+
+		var user User
+
+		bufferedReader := bufio.NewReader(r)
+		for {
+			line, err := bufferedReader.ReadBytes('\n')
+			if len(line) == 0 && err != nil {
+				if err != io.EOF {
+					errChannel <- err
+				}
+				break
 			}
-			return result, err
-		}
 
-		if err = json.Unmarshal(line, &user); err != nil {
-			return result, err
-		}
+			if err = json.Unmarshal(line, &user); err != nil {
+				errChannel <- err
+				break
+			}
 
-		result = append(result, user)
-	}
+			usersChannel <- user
+		}
+	}()
+
+	return usersChannel
 }
 
-func countDomains(u users, domain string) (DomainStat, error) {
+func countDomains(usersChannel <-chan User, errChannel <-chan error, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 
-	for _, user := range u {
-		if strings.Contains(user.Email, "."+domain) {
-			foundDomain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
-			result[foundDomain]++
+	for {
+		select {
+		case user, ok := <-usersChannel:
+			if !ok {
+				return result, nil
+			}
+			if strings.Contains(user.Email, "."+domain) {
+				foundDomain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+				result[foundDomain]++
+			}
+		case err := <-errChannel:
+			return nil, err
 		}
 	}
-	return result, nil
 }