--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -4,12 +4,13 @@ package hw10programoptimization
 
 import (
 	"bytes"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 )
 
-func TestGetDomainStat(t *testing.T) {
+func TestGetDomainStatAcceptance(t *testing.T) {
 	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
 {"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
@@ -36,4 +37,64 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+}
+
+func TestGetDomainStat(t *testing.T) {
+	tests := []struct {
+		name    string
+		data    string
+		domain  string
+		want    DomainStat
+		wantErr bool
+	}{
+		{
+			"user has domain email",
+			`{"Email":"aliquid_qui_ea@Browsedrive.gov"}`,
+			"gov",
+			DomainStat{"browsedrive.gov": 1},
+			false,
+		},
+		{
+			"two users has domain email",
+			`{"Email":"aliquid_qui_ea@Browsedrive.gov"}
+{"Email":"RoseSmith@browseDrive.gov"}`,
+			"gov",
+			DomainStat{"browsedrive.gov": 2},
+			false,
+		},
+		{
+			"all emails",
+			`{"Email":"aliquid_qui_ea@Browsedrive.gov"}
+{"Email":"mLynch@broWsecat.com"}`,
+			"",
+			DomainStat{"browsedrive.gov": 1, "browsecat.com": 1},
+			false,
+		},
+		{
+			"user hasn't domain email",
+			`{}`,
+			"gov",
+			DomainStat{},
+			false,
+		},
+		{
+			"returns parse error",
+			`{"Email":"aliquid_qui_ea@Browsedrive.gov",}`,
+			"gov",
+			DomainStat{},
+			true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			reader := strings.NewReader(tt.data)
+			got, err := GetDomainStat(reader, tt.domain)
+			if tt.wantErr {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+				require.Equal(t, tt.want, got)
+			}
+		})
+	}
 }