@AlexandrNikitin3776 2022-11-06
Результат benchstat

|name             |old time/op    |new time/op    |delta|
|-|-|-|-|
|GetDomainStat-8     |359ms ± 0%     |169ms ± 0%  |-52.77%|

|name             |old alloc/op  | new alloc/op |  delta|
|-|-|-|-|
|GetDomainStat-8 |    322MB ± 0% |     30MB ± 0% | -90.84%|

|name |            old allocs/op | new allocs/op |  delta|
|-|-|-|-|
|GetDomainStat-8 |    3.15M ± 0%  |   0.72M ± 0%  | -76.98% |